+++ airbyte-integrations/connectors/source-faker/source_faker/airbyte_message_with_cached_json.py
@@ -6,7 +6,12 @@ from airbyte_cdk.models import AirbyteMessage
 
 
 class AirbyteMessageWithCachedJSON(AirbyteMessage):
-    """I a monkeypatch to AirbyteMessage which pre-renders the JSON-representation of the object upon initialization.  This allows the JSON to be calculated in the thread/process that builds the object rather than the main thread."""
+    """
+    I a monkeypatch to AirbyteMessage which pre-renders the JSON-representation of the object upon initialization.
+    This allows the JSON to be calculated in the thread/process that builds the object rather than the main thread.
+
+    Note: We can't use @cache here because the LRU cache is not serializable when passed to child workers.
+    """
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
